\name{vote-package}
\alias{vote-package}
\alias{vote}
\docType{package}
\title{
Vote Counting
}
\description{
Counting votes via different methods, including single transferable votes or approval voting.
}
\details{


}
\author{
Hana Sevcikova, Bernard Silverman, Adrian Raftery

Maintainer: Hana Sevcikova
}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
